batch test runner: added optimization which goes through already run test batches much faster so we have a faster restart of later images' batch tests for an improved tester experience.  (That would be *me*.  ;-) )

+++ sizes/run_tess_B-round.sh
@@ -55,9 +55,12 @@ for IMG in DERIVSRC-* ; do
         for OEM in       3 ; do
             mkdir OEM${OEM}
             pushd OEM${OEM}
+			REDUCE=0
             for PSM in  6     1 3 11 13 ; do
                 for THRESH in 1 0 2 ; do
-                    cat > PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh  <<EOT
+					if ! test -f PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh ; then
+						REDUCE=1
+						cat > PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh  <<EOT
 if ! test -f PSM${PSM}-TH${THRESH}-${SIZE}-debug-2.log ; then
 (
 	# sometimes a tesseract run hangs; haven't found a decent clue why, so we apply a fixed timeout/abort to keep the run going, no matter what happens.
@@ -65,12 +68,14 @@ if ! test -f PSM${PSM}-TH${THRESH}-${SIZE}-debug-2.log ; then
 ) &
 fi
 EOT
-                    cat ./PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh
-                    ./PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh
+						cat ./PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh
+						./PSM${PSM}-TH${THRESH}-${SIZE}-cmdline.sh
+					fi
                 done
             done
 
-            cat <<EOT
+			if test ${REDUCE} != 0 ; then
+				cat <<EOT
 
 
 ######################################################################################
@@ -93,24 +98,25 @@ EOT
 
 EOT
 
-            echo "Waiting for the tesseract runs to finish..."
-            while true ; do
-                if test $( ps ax | grep -e tesseract | wc -l ) -le 2 ; then
-                    break
-                fi
-                echo "sleep..."
-                sleep 1
-            done
+				echo "Waiting for the tesseract runs to finish..."
+				while true ; do
+					if test $( ps ax | grep -e tesseract | wc -l ) -le 2 ; then
+						break
+					fi
+					echo "sleep..."
+					sleep 1
+				done
 
-            echo "Reducing log files to reasonable size; only keeping their tails..."
-            for f in $( find . -name '*.log' -type f -mmin +15 -size +1M -print ) ; do
-                if test -f $f ; then
-                    tail -n  8000 $f > $f.8Klines-reduced
-                    rm $f
-                fi
-            done
+				echo "Reducing log files to reasonable size; only keeping their tails..."
+				for f in $( find . -name '*.log' -type f -mmin +15 -size +1M -print ) ; do
+					if test -f $f ; then
+						tail -n  8000 $f > $f.8Klines-reduced
+						rm $f
+					fi
+				done
 
-            rm *.pdf
+				rm *.pdf
+			fi
 
             popd
         done
